Safari-609.1.11.1/LayoutTests/http/tests/appcache/local-content-expected.txt

CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: line 28: Not allowed to load local resource: stdio.h
Test that documents loaded from application cache don't get access to local resources accidentally.

Should say SUCCESS:

SUCCESS
